requirements.txt
django
django-classy-tags==0.6.1
